Move EICode to utils
