+++ loopflow-computation/src/test/java/com/farao_community/farao/loopflow_computation/LoopFlowComputationTest.java
@@ -10,7 +10,7 @@ import com.farao_community.farao.data.crac_api.Cnec;
 import com.farao_community.farao.data.crac_api.Crac;
 import com.farao_community.farao.data.crac_loopflow_extension.CnecLoopFlowExtension;
 import com.farao_community.farao.data.crac_loopflow_extension.CracLoopFlowExtension;
-import com.farao_community.farao.data.glsk.import_.EICode;
+import com.farao_community.farao.util.EICode;
 import com.farao_community.farao.flowbased_computation.glsk_provider.GlskProvider;
 import com.farao_community.farao.flowbased_computation.glsk_provider.UcteGlskProvider;
 import com.farao_community.farao.util.LoadFlowService;
